[feat] RoomList, Filter 수정~~
